anthropic-python-sdk/pyproject.toml
[project]
name = "anthropic-python-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agentops>=0.3.21",
    "anthropic>=0.42.0",
    "packaging>=24.2",
    "python-dotenv>=1.0.1",
]
